+++ apps/client/src/pages/time-table/hooks/use-get-festival-to-add.ts
@@ -1,5 +1,5 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
-import { GET_FESTIVAL_TO_ADD_QUERY_OPTIONS } from '@shared/apis/time-table/get-festival-to-add-queries';
+import { GET_FESTIVAL_TO_ADD_QUERY_OPTIONS } from '@shared/apis/confeti-detail/get-festival-to-add-queries';
 
 export const useGetFestivalToAdd = () => {
   const { data, fetchNextPage, hasNextPage } = useInfiniteQuery({
