--- apps/client/src/shared/apis/confeti-detail/concert.ts
@@ -1,13 +1,0 @@
-import { BaseResponse } from '@shared/types/api';
-import { ConcertDetailResponse } from '@shared/types/concert-response';
-import { END_POINT } from '@shared/constants/api';
-import { get } from '../config/instance';
-
-export const getConcertDetail = async (
-  concertId: number,
-): Promise<ConcertDetailResponse> => {
-  const response = await get<BaseResponse<ConcertDetailResponse>>(
-    `${END_POINT.GET_CONCERT_DETAIL}/${concertId}`,
-  );
-  return response.data;
-};
