+++ apps/client/src/shared/apis/confeti-detail/confeti.ts
@@ -1,7 +1,17 @@
 import { BaseResponse } from '@shared/types/api';
+import { ConcertDetailResponse } from '@shared/types/concert-response';
 import { FestivalDetailResponse } from '@shared/types/festival-response';
 import { END_POINT } from '@shared/constants/api';
 import { get } from '../config/instance';
+
+export const getConcertDetail = async (
+  concertId: number,
+): Promise<ConcertDetailResponse> => {
+  const response = await get<BaseResponse<ConcertDetailResponse>>(
+    `${END_POINT.GET_CONCERT_DETAIL}/${concertId}`,
+  );
+  return response.data;
+};
 
 export const getFestivalDetail = async (
   festivalId: number,
