+++ apps/client/src/shared/apis/confeti-detail/get-festival-to-add-queries.ts
