+++ apps/client/src/shared/apis/confeti-detail/performance-queries.ts
@@ -1,6 +1,8 @@
-import { queryOptions } from '@tanstack/react-query';
-import { getConcertDetail } from './concert';
-import { getFestivalDetail } from './festival';
+import { infiniteQueryOptions, queryOptions } from '@tanstack/react-query';
+import { getConcertDetail, getFestivalDetail } from './confeti';
+import { getFestivalToAdd } from './get-festival-to-add';
+
+import { GetFestivalToAddResponse } from '@shared/types/get-festival-to-add-response';
 
 export const PERFORMANCE_QUERY_KEY = {
   ALL: ['performances'],
@@ -14,6 +16,10 @@ export const PERFORMANCE_QUERY_KEY = {
     'festival',
     festivalId,
   ],
+  GET_FESTIVAL_TO_ADD: {
+    ALL: ['getFestivalToAdd'],
+    LIST: () => [...PERFORMANCE_QUERY_KEY.GET_FESTIVAL_TO_ADD.ALL, 'list'],
+  },
 } as const;
 
 export const PERFORMANCE_QUERY_OPTIONS = {
@@ -26,5 +32,12 @@ export const PERFORMANCE_QUERY_OPTIONS = {
     queryOptions({
       queryKey: PERFORMANCE_QUERY_KEY.FESTIVAL(festivalId),
       queryFn: () => getFestivalDetail(festivalId),
+    }),
+  GET_FESTIVAL_TO_ADD_LIST: () =>
+    infiniteQueryOptions<GetFestivalToAddResponse, Error>({
+      queryKey: PERFORMANCE_QUERY_KEY.GET_FESTIVAL_TO_ADD.LIST(),
+      queryFn: ({ pageParam }) => getFestivalToAdd(pageParam as number),
+      initialPageParam: undefined,
+      getNextPageParam: (lastPage) => lastPage.nextCursor || undefined,
     }),
 };
